feat: create image request hook and sticky prompt action panel
