+++ app/create/page.tsx
@@ -5,6 +5,7 @@ import PromptInputForm from './_component/PromptInputForm'
 import PromptActionPanel from './_component/PromptActionPanel'
 import PendingImagesPanel from '../_components/PendingImagesPanel'
 import AdvancedOptions from '../_components/AdvancedOptions'
+import PromptStickyCreate from './_component/PromptStickyCreate'
 
 export const metadata: Metadata = {
   title: 'Create | ArtBot for Stable Diffusion'
@@ -15,9 +16,9 @@ export const metadata: Metadata = {
 export default async function CreatePage() {
   return (
     <PromptInputProvider>
-      <PageTitle>Create</PageTitle>
       <div className="col md:row justify-start !items-start !gap-4">
-        <div className="sm:min-w-[448px] w-full md:max-w-[512px] col gap-4">
+        <div className="sm:min-w-[448px] w-full md:max-w-[512px] col gap-2">
+          <PageTitle>Create</PageTitle>
           <PromptInputForm />
           <PromptActionPanel />
           <AdvancedOptions />
@@ -28,7 +29,10 @@ export default async function CreatePage() {
             minHeight: 'calc(100vh - 56px)'
           }}
         >
-          <PendingImagesPanel />
+          <div className="col w-full">
+            <PromptStickyCreate />
+            <PendingImagesPanel />
+          </div>
         </div>
       </div>
     </PromptInputProvider>
